Updated player-enemy collision detection- computes quick heuristic
Previously collision detection was computing magnitude of the 2D distance between the player and each enemy each frame. This was an inefficient operation to be doing for each enemy each frame. 

To check the collision is possible, first check the x distance makes a collision possible- this is a simple subtraction. Only compute magnitude if this calculation finds a collision is possible.

This approach can be applied in other parts of the code where distances are being computed each frame.