Added similar mechanism to stop player moving partway into walls (left or right). Moved functions from Level to Character
Adjust the x position to ensure the character can't end up part-way through the wall during the frame.
This functionality belongs more in Character, over Level
Put this in Character rather than Player, to allow it to be reused for enemies. Although it will be most useful for players due to their movement.