Updated light bolts- they can no longer travel through walls, floors etc.
It didn't really make sense for light to travel through walls, so stopped this behaviour! The bolt checks each cell it crosses as it's being drawn, and determines if the cell its entering is a wall/floor, in which case it sets that as the end point.
There's probably some scope for optimisation here. Can at the very least check less frequently. Or do something with division to achieve this