Firing a bolt now causes light radius to decrease smoothly
Previously firing a bolt would cause the light radius to immediately decrease by 100 pixels, which was quite jarring. This change replaces this with increasing the rate at which the radius decreases, while a bolt is being fired. This results in this looking much smoother, while still clearly showing that firing a bolt discharges the lantern