Removing magic numbers
Planning to make some large improvements to program structure fairly soon. A lot of things that should really be in Level are instead in Main, which compounds as it requires new things to go in Main also.
Emphasis on improving code quality to allow that to go as smoothly as possible.