More code out of Main- enemy drawing
Enemies were being drawn conditionally in Main, depending on enemy type, whether it was alive, direction, step count etc. Although this was ultimately going to an enemy draw function, all the actual work to figure out what to draw was being done by Main. I moved all this code into the enemy classes, and no longer pass the image as a parameter as this can be accessed directly from Main