Made checking for falling out of the level more efficient, added new PVector for character cell position
When testing on a lower powered PC I noticed performance stutters at times. This, combined with the fact I know there's a lot of inefficient code in this project, has made me prioritise this for the time being.

Previously we just had a PVector for the character pixel coordinates. While this works, we do a lot of calculations each frame that require conversion of pixel location -> cell. Storing the cell coordinates allows for a lot of these conversions to be removed- it can be calculated just once per frame for each character. It may even be possible to infer if the character has crossed a cell boundary and only perform the calculation then