Optimisation: Using the new row/col mappings in collision detection
We check for collisions each frame, for the player, every enemy and bullet, for each of the four directions. So A LOT 
Previously we were getting the pixel values for the start of each cell by multiplying row/col with the cell width/height. Using the parallel arrays I set up in my last commit, this is now done in constant time