Bugfixes and optimisation around getting cell coordinates
Bugfix 1: Player was getting initialised within the wall
Bugfix 2: Player couldn't approach a wall to the right
Optimised getting cell coordinates from pixel values, by using the parallel arrays the other way around in a for loop. Previously we were using multiplication for this, and this change reduces the complexity to linear. Currently this function is only used for the bolt, but there's scope to use this more widely in the future