Minor refactoring
Spent time thinking about how the bolt is drawn, and whether it's too inefficient.
The collision checks were previously rather inefficient, but are much better following recent changes. Currently it just calculates the unit vector and the full bolt vector each frame, before checking for collisions at regular intervals.
While it may be possible to be smarter about collisions by calculating exactly where the grid boundaries are, we'd still need to calculate the unit vectors etc., I'm not convinced this would be any more efficient than what we currently do, as there'd inevitably be some overhead involved when calculating these boundaries. So for now I'll leave this code as-is